+++ NEWS.md
@@ -5,6 +5,10 @@
 
 # mlr3extralearners 0.6.0-9000
 
+* BREAKING CHANGE: lightgbm's early stopping mechanism now uses the task's test set.
+* Add two new learners `regr.abess` and `classif.abess` (thanks to @bbayukari)
+
+
 # mlr3extralearners 0.6.0
 
 * Feat: Added learner `LearnerClassifGlmer` (https://github.com/mlr-org/mlr3extralearners/issues/243)